requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1

celery~=5.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1